Fix AIS 6:1:3 and add testing.
Added spare3 and spare4 to handle the large spare block at the end of the message.
The authors of 6:1:3 were nuts with the giant spare block.  That was a
terrible design decision.  The message is often used in a short form
that violates the spec.

Added support for unsigned ints in the python interface.
Debated int versus unsigned int versus uint32_t from cstdint.  I'm not
sure which is best.  Went with unsigned int for right now.