Prototype safe wrapper for AisBitset accesses
Provides safe a batch-fetch interface for AisBitset bit sub-sequences,
erroring on out-of-bounds requests at runtime or dying on an assert if
they're enabled.

The interface is supposed to match libais's current control flow
approach, avoiding requiring major code surgery to integrate. It's
currently a bit too wordy, though, and could be improved.

This is designed to tackle a semi-regular class of bugs in libais,
where a decoder-set message length precondition is narrower than the
AisBitset calls made further down the function. Since asserts are the
only error handling in this situation we currently get UB in release
builds.