ais_py.cpp: Prefer PyDict_SetItemString over PyDict_SetItem
All insertions into python dictionnaries use a string as key and used to
do it with PyDict_SetItem.  This function requires a PyObject* as
argument, when a variant exists that only require a const char*.  Using
the latter function avoids the creation of a intermediary key object.

Moreover, in few places the temporary object where not cleared properly:

   PyDict_SetItem(dict, PyUnicode_FromString("cargos"), cargo_list);

This allocates a PyObject* from the "cargos" string, and gives it
directly to PyDict_SetItem.  The problem is that the refcount of the
temporary after PyUnicode_FromString is 1, it is then incremented by
PyDict_SetItem to 2. When the containing dict is released, this key is
kept leaking because Py_DECREF was not called on the temporary.

The correct usage would have been to use:

   PyObject *key = PyUnicode_FromString("cargos");
   assert(key);
   PyDict_SetItem(dict, key, cargo_list);
   Py_DECREF(key);

Hopefully, PyDict_SetItemString factores this out so we can just use the
following:

   PyDict_SetItemString(dict, "cargos", cargo_list);