ais_py.cpp: Py_DECREF lists once inserted in dicts.
Along the creation of python dict, few lists are created and then
inserted into the dict using DictSafeSetItem.  This function internaly
uses PyDict_SetItemString which increments the refcount to the list.
Once this is done, a call to Py_DECREF should be made to indicate that
the c++ code does not own the list anymore.  Not doing so caused few
memory leaks.